pyproject.toml
[tool.poetry]
package-mode = false
name = "gptpython"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.35.8"
python-dotenv = "^1.0.1"
typechat = {git = "https://github.com/microsoft/TypeChat", subdirectory = "python"}
pandas = "^2.2.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
